pyproject.toml
[project]
name = "windchill-copilot-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.21",
    "matplotlib>=3.10.8",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "openai>=2.23.0",
    "pydot>=4.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
